Distribute as infrastructure-agent plugin.
Instead of trying to distribute this as a built-in Opsview plugin,
it makes a lot of sense to distribute it as an infrastructure-agent
plugin instead.

Instead of relying on setuid, use setcap to grant additional
capabilities to open raw icmp sockets.

Rename check-jitter-opsview to opsview-check-jitter to be in line with
other opsview package names.